Use TLC git repository for dependency
TLC has moved core development from SVN to Git. Currently added version
of TLC is equivalent to the trunk (r249) from SVN, with changes for
directory structure and a gitignore.

Also note that jscert is sensitive to TLC version changes (I've just
spent several hours trying to rebuild an old version with the wrong TLC
version...), so this is a win also in this respect.

Fixes #10