Force `make proof` to re-build Coq sourcefiles
`make` followed by `make proof` was failing, even though files had been
patched during the `make proof` step.

This solution is a bit hacky, I'd really like a Makefile command to
notify make that this build step changes these files. Setting the patch
command as a dependency on the source files will force them to be
patched each time. Order-only dependencies won't work as they trigger a
build only if the target doesn't exist (which is the exact opposite
situation of the one we have...)

Sigh.